Introduce base hierarchy to exceptions.
