+++ python/pyspark/sql/tests/test_dataframe.py
@@ -46,8 +46,6 @@ from pyspark.sql.types import (
 from pyspark.errors import (
     AnalysisException,
     IllegalArgumentException,
-    SparkConnectException,
-    SparkConnectAnalysisException,
     PySparkTypeError,
 )
 from pyspark.testing.sqlutils import (
@@ -948,8 +946,7 @@ class DataFrameTestsMixin:
         self.assertRaises(TypeError, lambda: self.spark.range(1).sample(seed="abc"))
 
         self.assertRaises(
-            (IllegalArgumentException, SparkConnectException),
-            lambda: self.spark.range(1).sample(-1.0).count(),
+            IllegalArgumentException, lambda: self.spark.range(1).sample(-1.0).count()
         )
 
     def test_toDF_with_schema_string(self):
@@ -1595,17 +1592,13 @@ class DataFrameTestsMixin:
         # incompatible field nullability
         schema4 = StructType([StructField("j", LongType(), False)])
         self.assertRaisesRegex(
-            (AnalysisException, SparkConnectAnalysisException),
-            "NULLABLE_COLUMN_OR_FIELD",
-            lambda: df.to(schema4).count(),
+            AnalysisException, "NULLABLE_COLUMN_OR_FIELD", lambda: df.to(schema4).count()
         )
 
         # field cannot upcast
         schema5 = StructType([StructField("i", LongType())])
         self.assertRaisesRegex(
-            (AnalysisException, SparkConnectAnalysisException),
-            "INVALID_COLUMN_OR_FIELD_DATA_TYPE",
-            lambda: df.to(schema5).count(),
+            AnalysisException, "INVALID_COLUMN_OR_FIELD_DATA_TYPE", lambda: df.to(schema5).count()
         )
 
     def test_repartition(self):
