+++ python/pyspark/sql/tests/test_functions.py
@@ -25,7 +25,7 @@ import math
 import unittest
 
 from py4j.protocol import Py4JJavaError
-from pyspark.errors import PySparkTypeError, PySparkValueError, SparkConnectException
+from pyspark.errors import PySparkTypeError, PySparkValueError
 from pyspark.sql import Row, Window, types
 from pyspark.sql.functions import (
     udf,
@@ -1056,6 +1056,9 @@ class FunctionsTestsMixin:
         self.assertEqual(date(2017, 1, 22), parse_result["to_date(dateCol)"])
 
     def test_assert_true(self):
+        self.check_assert_true(Py4JJavaError)
+
+    def check_assert_true(self, tpe):
         from pyspark.sql.functions import assert_true
 
         df = self.spark.range(3)
@@ -1065,10 +1068,10 @@ class FunctionsTestsMixin:
             [Row(val=None), Row(val=None), Row(val=None)],
         )
 
-        with self.assertRaisesRegex((Py4JJavaError, SparkConnectException), "too big"):
+        with self.assertRaisesRegex(tpe, "too big"):
             df.select(assert_true(df.id < 2, "too big")).toDF("val").collect()
 
-        with self.assertRaisesRegex((Py4JJavaError, SparkConnectException), "2000000"):
+        with self.assertRaisesRegex(tpe, "2000000"):
             df.select(assert_true(df.id < 2, df.id * 1e6)).toDF("val").collect()
 
         with self.assertRaises(PySparkTypeError) as pe:
@@ -1081,14 +1084,17 @@ class FunctionsTestsMixin:
         )
 
     def test_raise_error(self):
+        self.check_raise_error(Py4JJavaError)
+
+    def check_raise_error(self, tpe):
         from pyspark.sql.functions import raise_error
 
         df = self.spark.createDataFrame([Row(id="foobar")])
 
-        with self.assertRaisesRegex((Py4JJavaError, SparkConnectException), "foobar"):
+        with self.assertRaisesRegex(tpe, "foobar"):
             df.select(raise_error(df.id)).collect()
 
-        with self.assertRaisesRegex((Py4JJavaError, SparkConnectException), "barfoo"):
+        with self.assertRaisesRegex(tpe, "barfoo"):
             df.select(raise_error("barfoo")).collect()
 
         with self.assertRaises(PySparkTypeError) as pe:
